+++ R/cql-translator.R
@@ -10,6 +10,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
+#' @importFrom rlang :=
+
 # Function to translate R code to CQL
 cql_translate <- function(...) {
   ## convert dots to list of quosures
@@ -22,11 +24,28 @@ cql_translate <- function(...) {
   ## predicates and CQL() expressions are evaluated into valid CQL code
   ## so they can be combined with the rest of the query
   dots <- lapply(dots, function(x) {
-    rlang::new_quosure(
-      dbplyr::partial_eval(rlang::get_expr(x), env = rlang::get_env(x)),
-                rlang::get_env(x))
+
+    # make sure all arguments are named in the call so can be modified
+    x <- rlang::call_standardise(x, env = rlang::get_env(x))
+
+    # if an argument to a predicate is a function call, need to tell it to evaluate
+    # locally, as by default all functions are treated as remote and thus
+    # not evaluated. Do this by using `rlang::call2` to wrap the function call in
+    # local()
+    # See ?rlang::partial_eval and https://github.com/bcgov/bcdata/issues/146
+    for (call_arg in rlang::call_args_names(x)) {
+      if (is.call(rlang::call_args(x)[[call_arg]])) {
+        x <- rlang::call_modify(
+          x, !!call_arg := rlang::call2("local", rlang::call_args(x)[[call_arg]])
+        )
+      }
+    }
+
+    rlang::new_quosure(dbplyr::partial_eval(x), rlang::get_env(x))
   })
+
   sql_where <- dbplyr::translate_sql_(dots, con = cql_dummy_con, window = FALSE)
+
   build_where(sql_where)
 }