+++ tests/testthat/test-cql-string.R
@@ -41,19 +41,19 @@ test_that("All cql geom predicate functions work", {
   }
   expect_equal(
     DWITHIN(the_geom, 1), #default units meters
-    CQL("DWITHIN({geom_name}, POINT (1 1), 1, 'meters')")
+    CQL("DWITHIN({geom_name}, POINT (1 1), 1, meters)")
   )
   expect_equal(
     DWITHIN(the_geom, 1, "meters"),
-    CQL("DWITHIN({geom_name}, POINT (1 1), 1, 'meters')")
+    CQL("DWITHIN({geom_name}, POINT (1 1), 1, meters)")
   )
   expect_equal(
     BEYOND(the_geom, 1, "feet"),
-    CQL("BEYOND({geom_name}, POINT (1 1), 1, 'feet')")
+    CQL("BEYOND({geom_name}, POINT (1 1), 1, feet)")
   )
   expect_equal(
     RELATE(the_geom, "*********"),
-    CQL("RELATE({geom_name}, POINT (1 1), '*********')")
+    CQL("RELATE({geom_name}, POINT (1 1), *********)")
   )
   expect_equal(
     BBOX(c(1,2,1,2)),